requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-bcrypt==1.0.1
pymongo==4.5.0
mongoengine==0.24.2
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1